chore: refactor available destinations to be calculated on the filter component
