+++ packages/frontend/src/components/SchedulersView/SchedulersTable.tsx
@@ -51,12 +51,7 @@ import {
 } from 'react';
 import { Link, useSearchParams } from 'react-router';
 import { usePaginatedSchedulers } from '../../features/scheduler/hooks/useScheduler';
-import {
-    useSchedulerFilters,
-    type DestinationType,
-} from '../../features/scheduler/hooks/useSchedulerFilters';
-import useHealth from '../../hooks/health/useHealth';
-import { useGetSlack } from '../../hooks/slack/useSlack';
+import { useSchedulerFilters } from '../../features/scheduler/hooks/useSchedulerFilters';
 import { useIsTruncated } from '../../hooks/useIsTruncated';
 import { useProject } from '../../hooks/useProject';
 import GSheetsSvg from '../../svgs/google-sheets.svg?react';
@@ -242,25 +237,6 @@ const SchedulersTable: FC<SchedulersTableProps> = ({
     useEffect(() => {
         fetchMoreOnBottomReached(tableContainerRef.current);
     }, [fetchMoreOnBottomReached]);
-
-    const health = useHealth();
-    const slack = useGetSlack();
-    const organizationHasSlack = !!slack.data?.organizationUuid;
-
-    // Compute available destinations based on integrations
-    const availableDestinations = useMemo<DestinationType[]>(() => {
-        const destinations: DestinationType[] = [];
-        if (health.data?.hasEmailClient) {
-            destinations.push('email');
-        }
-        if (organizationHasSlack) {
-            destinations.push('slack');
-        }
-        if (health.data?.hasMicrosoftTeams) {
-            destinations.push('msteams');
-        }
-        return destinations;
-    }, [health.data, organizationHasSlack]);
 
     const sorting = useMemo<MRT_SortingState>(
         () => [{ id: sortField, desc: sortDirection === 'desc' }],
@@ -842,7 +818,6 @@ const SchedulersTable: FC<SchedulersTableProps> = ({
                     currentResultsCount={totalFetched}
                     hasActiveFilters={hasActiveFilters}
                     onClearFilters={resetFilters}
-                    availableDestinations={availableDestinations}
                     selectedCount={selectedRows.length}
                     onBulkReassign={handleBulkReassign}
                 />
